Fix warning in distribution Makefile
Since distributions (as produced with "make dist") are not Git
repositories, when make parses the default Makefile included in the
distribution the "git describe" invocation used to infer the version for
"make dist" fails. This usually manifests with "make install", which is
the typical target used in a distribution, but all targets are
susceptible.

The problematic functionality is entirely specific to the "dist" target
and there is no meaningful way to resolve it: it's trivial to hard-code
the version but that would just leave the "dist" target to fail instead,
and that target doesn't even make sense in the context.

Solve this by dynamically stripping useless targets when generating the
distribution.

At this point, "git archive" seems increasingly like the wrong tool for
this job.