Strip credentials-part user from HTTPS URL
An HTTPS URL can include a user name in the credentials-part, before the
domain-part. With providers that support this, like Bitbucket, the user
only has to additionally supply the password when authenticating. URLs
like this tie into HTTP Basic authentication, however, so when browsers
access those URLs but are not asked to authenticate they can get
confused and ask the user if this is what they intended. It's better to
just strip the user part to eliminate this warning.